+++ storybook/src/example-pages/toepassingen/aanvragen/documenten.ts
@@ -1,19 +1,27 @@
 import { html } from "lit-html";
 import { examplePageFactory } from "../../../example-page-factory";
+import { footerPartial } from "../../partials/footer";
+import { headerPartial } from "../../partials/header";
+import { header } from "../../partials/header.content";
+import { mainMenu } from "./main-menu.content";
 
 examplePageFactory(
   "Voorbeeldpagina's",
   "Toepassingen/Aanvragen",
   "Documenten",
-  ({
-    accordionTemplate,
-    applicationHeadingTemplate,
-    buttonTemplate,
-    formButtonsTemplate,
-    formGroupFilesTemplate,
-    richContentTemplate,
-  }) => html`
+  (
+    {
+      accordionTemplate,
+      applicationHeadingTemplate,
+      buttonTemplate,
+      formButtonsTemplate,
+      formGroupFilesTemplate,
+      richContentTemplate,
+    },
+    templates,
+  ) => html`
     <div class="container">
+      ${headerPartial(templates, { ...header, mainMenu })}
       <main>
         <form>
           ${applicationHeadingTemplate({
@@ -142,6 +150,7 @@ examplePageFactory(
             ],
           })}
         </form>
+        ${footerPartial(templates)}
       </main>
     </div>
   `,