+++ storybook/src/example-pages/toepassingen/aanvragen/landingspagina.ts
@@ -4,6 +4,7 @@ import { examplePageFactory } from "../../../example-page-factory";
 import { footerPartial } from "../../partials/footer";
 import { headerPartial } from "../../partials/header";
 import { header } from "../../partials/header.content";
+import { mainMenu } from "./main-menu.content";
 
 examplePageFactory(
   "Voorbeeldpagina's",
@@ -11,32 +12,7 @@ examplePageFactory(
   "Landingspagina",
   ({ anchorTemplate, buttonTemplate, highlightBoxTemplate, richContentTemplate }, templates) => html`
     <div class="container">
-      ${headerPartial(templates, {
-        ...header,
-        mainMenu: [
-          {
-            label: "Home",
-            url: "#",
-          },
-          {
-            label: "Vergunningscheck",
-            url: "#",
-          },
-          {
-            label: "Aanvragen",
-            url: "#",
-            active: true,
-          },
-          {
-            label: "Regels op de kaart",
-            url: "#",
-          },
-          {
-            label: "Mijn Omgevingsloket",
-            url: "#",
-          },
-        ],
-      })}
+      ${headerPartial(templates, { ...header, mainMenu })}
       <main>
         <div
           class="row dso-banner dso-banner-implementation-specific-image"