+++ storybook/src/example-pages/toepassingen/aanvragen/project.ts
@@ -2,13 +2,18 @@ import { html } from "lit-html";
 
 import { examplePageFactory } from "../../../example-page-factory";
 import { form } from "./project.content";
+import { mainMenu } from "./main-menu.content";
+import { headerPartial } from "../../partials/header";
+import { footerPartial } from "../../partials/footer";
+import { header } from "../../partials/header.content";
 
 examplePageFactory(
   "Voorbeeldpagina's",
   "Toepassingen/Aanvragen",
   "Project",
-  ({ applicationHeadingTemplate, alertTemplate, formTemplate, formButtonsTemplate, anchorTemplate }) => html`
+  ({ applicationHeadingTemplate, alertTemplate, formTemplate, formButtonsTemplate, anchorTemplate }, templates) => html`
     <div class="container">
+      ${headerPartial(templates, { ...header, mainMenu })}
       <main>
         ${applicationHeadingTemplate({ title: "Aanvragen", subtitle: "1. Project", step: "Stap 1/7" })}
         ${alertTemplate({
@@ -29,6 +34,7 @@ examplePageFactory(
           ],
         })}
       </main>
+      ${footerPartial(templates)}
     </div>
   `,
 );